requirements.txt
automationhat==0.1.0
RPi.GPIO==0.6.5
APScheduler
